refactor: put entity filter from the csr info in CSR class
